feat:Deprecate createGlobalState in favor of createSharedState
