+++ docs/createSharedState.md
@@ -1,29 +1,29 @@
-# `useGlobalState`
+# `useSharedState`
 
 > :warning:
-> This function is deprecated in favor of [`useSharedState`[(useSharedState.md)] since it doesn't create a trully global state but a shared one.
+> This function is deprecated in favor of [`useSharedState`[(useSharedState.md)] since it doesn't create a trully Shared state but a shared one.
 
-A React hook that creates a globally shared state.
+A React hook that creates a shared state.
 
 ## Usage
 
 ```tsx
-const useGlobalValue = createGlobalState<number>(0);
+const useSharedValue = createSharedState<number>(0);
 
 const CompA: FC = () => {
-  const [value, setValue] = useGlobalValue();
+  const [value, setValue] = useSharedValue();
 
   return <button onClick={() => setValue(value + 1)}>+</button>;
 };
 
 const CompB: FC = () => {
-  const [value, setValue] = useGlobalValue();
+  const [value, setValue] = useSharedValue();
 
   return <button onClick={() => setValue(value - 1)}>-</button>;
 };
 
 const Demo: FC = () => {
-  const [value] = useGlobalValue();
+  const [value] = useSharedValue();
   return (
     <div>
       <p>{value}</p>
@@ -37,22 +37,22 @@ const Demo: FC = () => {
 It also allows initializing the state with a function and using a function in the setState:
 
 ```tsx
-const useGlobalValue = createGlobalState<number>(() => 0);
+const useSharedValue = createSharedState<number>(() => 0);
 
 const CompA: FC = () => {
-  const [value, setValue] = useGlobalValue();
+  const [value, setValue] = useSharedValue();
 
   return <button onClick={() => setValue(value => value + 1)}>+</button>;
 };
 
 const CompB: FC = () => {
-  const [value, setValue] = useGlobalValue();
+  const [value, setValue] = useSharedValue();
 
   return <button onClick={() => setValue(value => value - 1)}>-</button>;
 };
 
 const Demo: FC = () => {
-  const [value] = useGlobalValue();
+  const [value] = useSharedValue();
   return (
     <div>
       <p>{value}</p>
