+++ src/index.ts
@@ -113,4 +113,5 @@ export { useRendersCount } from './useRendersCount';
 export { useFirstMountState } from './useFirstMountState';
 export { default as useSet } from './useSet';
 export { createGlobalState } from './factory/createGlobalState';
+export { createSharedState } from './factory/createSharedState';
 export { useHash } from './useHash';
